search: factor out callSearcherOverRepos
This closure doesn't need to live in searchFilesInRepos anymore. We can
now pass in the search stream. The function now has its own trace,
waitgroup and error reporting.

Previously we relied on setError, which records the first error. We also
relied on waitgroup. We still as a whole wait on the same waitgroup and
will report the first error. It is now done via an errgroup which is
created for callSearcherOverRepos. This simplifies error reporting and
cancellation.